0.0.0.tipsAgents/src/agente_roteador/pyproject.toml
[tool.poetry]
name = "agente-roteador"
version = "0.1.0"
description = "Agente Roteador"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.13"
pyyaml = "^6.0"
google-generativeai = "^0.3.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
cryptography = "^42.0.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"  # Added for FastAPI TestClient
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
pythonpath = ["."]
testpaths = ["test"]
asyncio_mode = "auto"
addopts = "--basetemp=test/fixtures"
